freqresp: aligning with the built-in version
Freqresp now returns a column vector of frequencies and accepts imaginary and complex frequencies, just like in the built-in case.
If all frequencies are real, then the transfer function is evaluated over the imaginary axis at those frequencies, otherwise on the complex plane over the frequencies specified by the user.