Major benchmarksCheck update for better usability
The output of benchmarkChecks was updated.

benchmarksCheck is now able to download the benchmarks from https://webdisk.ads.mwn.de

benchmarksCheck now throws an error in case there are problems when
downloading/copying files from source to destination.

The inputs of benchmarksCheck where changed from optional to
parameters in order to improve usability.