bodeToFig - changed function to copy only visible axes
when using bodeToFig to a plot generated using bodemag, the phase (which is not visible) was copied as well. Using matlab2tikz on the bodeToFig converted figure would therefore plot the phase as well.

To avoid this, bodeToFig was adapted to copy only visible axes, creating a new figure from the visible part of the current one