frd - created function frd to replace freqresp(...,struct('frd',1)) usage
the extra functionality added to freqresp deviated from the built-in. As such, to create an frd object of sss and ss objects, you need two different calls. To avoid this, the function frd was created to be used just like in the built-in case: sysfrd = frd(sys,w).

It can be even used without specifying a frequency vector w.

All other functions with dependencies were adapted accordingly, all sss tests passed.